Potecting Edit and Delete in Backend

+++ backend/routes/posts.js
@@ -65,9 +65,12 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
     //imagePath: url + "/images/" + req.file.filename
 
   });
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: 'Post Updated!'});
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message: 'Post Updated!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized to Update the post!'});
+    }
   });
 });
 
@@ -104,9 +107,12 @@ router.get('/:id',(req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: 'Post deleted!'});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: 'Post Deleted!'});
+    } else {
+      res.status(401).json({message: 'Not Authorized to Delete the post!'});
+    }
   });
 });
